backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid==3.3.4
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.9
Django==5.1.1
django-debug-toolbar==4.4.6
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.10
isort==5.13.2
kombu==5.4.2
mccabe==0.7.0
msgpack==1.1.0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
PyJWT==2.9.0
pylint==3.3.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.1.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
django-filter==24.3
